Split pets/__init__.py into modules.
